[css-filters] Migrate effect-reference-external-stylesheet to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters/ and
into the WPT-specific directory, adding links to the relevant specs and
a test assertion describing its purpose.

Additionally, rename svg-relative-urls-0001.html to
svg-relative-urls-001.html, and add this test as a second test for the
same use case.

Bug: 1063749
Change-Id: Ifc0a6abfa45a6e6ad16f08388210ddcb0286142a